now handles float

+++ pilot/Inc/Motors.h
@@ -4,6 +4,6 @@
 #include "stm32f4xx_hal.h"
 
 void Motors_Init(TIM_HandleTypeDef* htim);
-void Motors_Set(uint16_t* values);
+void Motors_Set(float* values);
 
 #endif
